% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProspectiveMain.R
\name{executeProspective}
\alias{executeProspective}
\title{Execute the Study}
\usage{
executeProspective(
  databaseDetails,
  siteId,
  outputFolder,
  cohortTable = "olglmm_covid",
  createCohorts = F,
  createData = F,
  createPlot = createData,
  sampleSize = NULL,
  run = F,
  verbosity = "INFO",
  dates = list(omicron_1 = list(startDate = "20211101", endDate = "20220228"), omicron_2
    = list(startDate = "20220301", endDate = "20220631"), omicron_3 = list(startDate =
    "20220701", endDate = "20221101"))
)
}
\arguments{
\item{databaseDetails}{The connection details and OMOP CDM details. Created using \code{PatientLevelPrediction::createDatabaseDetails}.}

\item{siteId}{The reference name used for your database site}

\item{outputFolder}{Name of local folder to place results - make sure control is in here; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{cohortTable}{The name of the table to create the cohorts}

\item{createCohorts}{Create the cohortTable table with the target population and outcome cohorts?}

\item{createData}{Create the labelled data set (this is required before runAnalysis)}

\item{createPlot}{Plot the number of covid hospitalizations by month}

\item{sampleSize}{The number of patients in the target cohort to sample (if NULL uses all patients)}

\item{run}{Runs the analysis - this require downloading the json control from https://pda-ota.pdamethods.org/ 
predictions.  This step requires downloading the updated json control from https://pda-ota.pdamethods.org/}

\item{verbosity}{Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:
\itemize{
\item{DEBUG}{Highest verbosity showing all debug statements}
\item{TRACE}{Showing information about start and end of steps}
\item{INFO}{Show informative information (Default)}
\item{WARN}{Show warning messages}
\item{ERROR}{Show error messages}
\item{FATAL}{Be silent except for fatal errors}
}}

\item{dates}{A named list containing the start and end dates for each wave investigated}
}
\description{
Execute the Study
}
\details{
This function executes the OLGLMM-COVID Study.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")

execute(databaseDetails,
        siteId = 'db1',
        outputFolder = "c:/temp/study_results", 
        createCohorts = T,
        createData = T,
        sampleSize = 10000,
        run = F,
        verbosity = "INFO"
        )
}

}
